+++ changelog.md
@@ -7,6 +7,7 @@
 ### 2021-01-21 9.2.0 ###
 
 * Fix: Compatibility with Astor Theme.
+* Improved: Added notes that Shortcode UI is not maintained to the button and readme.
 * Improved: Classic Tabs, new settings section for upcoming [ARVE sticky Videos](https://nextgenthemes.com/plugins/arve-sticky-videos/).
 * Improved: Lots of code restructured and improved.